packages/cli: move plugin diff prompts to separate file
