+++ packages/cli/src/commands/plugin/diff/index.ts
@@ -14,12 +14,14 @@
  * limitations under the License.
  */
 
-import chalk from 'chalk';
-import inquirer from 'inquirer';
 import { Command } from 'commander';
 import { readTemplateFiles } from './read';
 import { handlers, handleAllFiles } from './handlers';
-import { PromptFunc } from './types';
+import {
+  inquirerPromptFunc,
+  makeCheckPromptFunc,
+  yesPromptFunc,
+} from './prompts';
 
 const fileHandlers = [
   {
@@ -41,46 +43,12 @@ const fileHandlers = [
   },
 ];
 
-const inquirerPromptFunc: PromptFunc = async (msg) => {
-  const { result } = await inquirer.prompt({
-    type: 'confirm',
-    name: 'result',
-    message: chalk.blue(msg),
-  });
-  return result;
-};
-
-const makeCheck = () => {
-  let failed = false;
-
-  const promptFunc: PromptFunc = async (msg) => {
-    failed = true;
-    console.log(chalk.red(`[Check Failed] ${msg}`));
-    return false;
-  };
-
-  const finalize = () => {
-    if (failed) {
-      throw new Error(
-        'Check failed, the plugin is not in sync with the latest template',
-      );
-    }
-  };
-
-  return [promptFunc, finalize] as const;
-};
-
-const yesPromptFunc: PromptFunc = async (msg) => {
-  console.log(`Accepting: "${msg}"`);
-  return true;
-};
-
 export default async (cmd: Command) => {
   let promptFunc = inquirerPromptFunc;
   let finalize = () => {};
 
   if (cmd.check) {
-    [promptFunc, finalize] = makeCheck();
+    [promptFunc, finalize] = makeCheckPromptFunc();
   } else if (cmd.yes) {
     promptFunc = yesPromptFunc;
   }
